Add conftest.py, consolidate fixtures
